+++ core/src/tangram.cpp
@@ -19,9 +19,8 @@
 #include "scene/lights.h"
 #include "util/error.h"
 #include "stl_util.hpp"
-#include "util/textureCube.h"
-#include "glm/gtc/type_ptr.hpp"
 #include "util/tileID.h"
+#include "util/skybox.h"
 
 namespace Tangram {
 
@@ -31,18 +30,8 @@ namespace Tangram {
     std::shared_ptr<LabelContainer> m_labelContainer;
     std::shared_ptr<FontContext> m_ftContext;
     std::shared_ptr<DebugStyle> m_debugStyle;
+    std::shared_ptr<Skybox> m_skybox;
     
-    struct PosVertex {
-        // Position Data
-        GLfloat pos_x;
-        GLfloat pos_y;
-        GLfloat pos_z;
-    };
-    
-    std::shared_ptr<ShaderProgram> m_skyboxShader;
-    std::shared_ptr<Texture> m_skyboxTexture;
-    std::shared_ptr<TypedMesh<PosVertex>> m_skyboxMesh;
-
     static float g_time = 0.0;
     static unsigned long g_flags = 0;
 
@@ -108,44 +97,8 @@ namespace Tangram {
             std::unique_ptr<DebugStyle> debugStyle(new DebugStyle("Debug"));
             m_scene->addStyle(std::move(debugStyle));
             
-            // Skybox test
-            {
-                std::string fragShaderSrcStr = stringFromResource("cubemap.fs");
-                std::string vertShaderSrcStr = stringFromResource("cubemap.vs");
-            
-                m_skyboxShader = std::make_shared<ShaderProgram>();
-                m_skyboxShader->setSourceStrings(fragShaderSrcStr, vertShaderSrcStr);
-            
-                m_skyboxTexture = std::shared_ptr<Texture>(new TextureCube("cubemap.png"));
-                auto layout = std::shared_ptr<VertexLayout>(new VertexLayout({
-                    {"a_position", 3, GL_FLOAT, false, 0},
-                }));
-                
-                m_skyboxMesh = std::shared_ptr<TypedMesh<PosVertex>>(new TypedMesh<PosVertex>(layout, GL_TRIANGLES));
-                
-                std::vector<int> indices = {
-                    5, 1, 3, 3, 7, 5, // +x
-                    6, 2, 0, 0, 4, 6, // -x
-                    2, 6, 7, 7, 3, 2, // +y
-                    5, 4, 0, 0, 1, 5, // -y
-                    0, 2, 3, 3, 1, 0, // +z
-                    7, 6, 4, 4, 5, 7  // -z
-                };
-                
-                std::vector<PosVertex> vertices = {
-                    { -1.0, -1.0,  1.0 },
-                    {  1.0, -1.0,  1.0 },
-                    { -1.0,  1.0,  1.0 },
-                    {  1.0,  1.0,  1.0 },
-                    { -1.0, -1.0, -1.0 },
-                    {  1.0, -1.0, -1.0 },
-                    { -1.0,  1.0, -1.0 },
-                    {  1.0,  1.0, -1.0 }
-                };
-                
-                m_skyboxMesh->addVertices(std::move(vertices), std::move(indices));
-                m_skyboxMesh->compileVertexBuffer();
-            }
+            m_skybox = std::shared_ptr<Skybox>(new Skybox("cubemap.png"));
+            m_skybox->init();
         }
 
         // Create a tileManager
@@ -262,16 +215,7 @@ namespace Tangram {
             style->teardown();
         }
 
-        // Skybox
-        {
-            m_skyboxTexture->bind(0);
-
-            glm::mat4 vp = m_view->getViewProjectionMatrix();
-
-            m_skyboxShader->setUniformMatrix4f("u_modelViewProj", glm::value_ptr(vp));
-            m_skyboxShader->setUniformi("u_tex", 0);
-            m_skyboxMesh->draw(m_skyboxShader);
-        }
+        m_skybox->draw(*m_view);
         
         while (Error::hadGlError("Tangram::render()")) {}
     }
